Fix (false) warning with gcc 9

+++ src/pevent_lib.cc
@@ -89,7 +89,8 @@ static DDRes pevent_register_cpu_0(const PerfWatcher *watcher, int watcher_idx,
     } else {
       LG_NFO("Failed to perf_event_open for watcher: %s - with attr.type=%s, "
              "exclude_kernel=%d",
-             watcher->desc, perf_type_str(attr.type), attr.exclude_kernel);
+             watcher->desc, perf_type_str(attr.type),
+             static_cast<int>(attr.exclude_kernel));
     }
   }
   // check if one of the configs was successful
